requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.4
black==20.8b1
click==7.1.2
Django==3.1.1
djangorestframework==3.12.2
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==7.2.0
pycodestyle==2.6.0
pytz==2020.1
regex==2020.10.28
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
